move init to addons
