+++ cli/src/enviroment.js
@@ -6,9 +6,8 @@ export const defaultConfig = {
     SERVER_PORT_UNSECURE: 8080,
     SERVER_PORT_SECURE: 4443,
     SERVER_NAME: 'localhost',
-    INSTALL_MERCURE: 1,
-    MERCURE_PUBLISHER_JWT_KEY: '', // defaulted by random key in setupQuestions
-    MERCURE_SUBSCRIBER_JWT_KEY: '',
+    MERCURE_ENABLE: '1',
+    MERCURE_JWT_KEY: '', // defaulted by random key in setupQuestions
     // for internal usage in the prompts    
     // APP_DIR: '../',
     INT_PORTS: true,
